Issue #16: Script engine and scripting prefix conversion and aliasing
InWorldz.Phlox is now Halcyon.Phlox.

As a consequence all `iw` prefixed scripting functions and `IW` prefixed constants have been converted to `hal` and `HAL` respectively.  This has been in a manner that preserves all scripting functionality of the old forms: the old InWorldz prefixes have been preserved but are deprecated for an indefinite time.